setvbuf / setlinebuf on Linux
Piping the output of 'dvbsnoop' to other processes on recent versions of Linux had some unexpected results, essentially the dvbsnoop output was buffed up to ~30 lines at a time before being flushed to the next process. This messed up my simple line-parser as all the lines arrived at the same time. This patch uses setvbuf/setlinebuf to disable line-buffering of the dvbsnoop output.
https://sourceforge.net/p/dvbsnoop/patches/2/ by https://sourceforge.net/u/userid-672436/profile/